Change to new trace format and make multiprocess tracing work.
Switch from the old {traceEvents: to the new [ format. trace_event_viewer
knows how to fixup traces that dont have a close } on them, allowing us to
remove complexity around adding trailing ]} to tracefiles.

When creating a trace file, output process metadata as the first action.
This allows us to make every flush command preceed itself with a ",", since
there is guaranteed to be a previous event. This allows more complexity to be
avoided.

Add _note function to help with debugging in the future.